김문경/requirements.txt
asgiref==3.5.2
Django==3.2.13
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
Markdown==3.4.1
PyJWT==1.7.1
pytz==2022.6
sqlparse==0.4.3
